--- a/src/System.CommandLine/Invocation/InvocationContext.cs
+++ b/src/System.CommandLine/Invocation/InvocationContext.cs
@@ -3,6 +3,7 @@
 
 using System.CommandLine.Binding;
 using System.CommandLine.Help;
+using System.CommandLine.IO;
 using System.CommandLine.Parsing;
 using System.Threading;
 
@@ -16,6 +17,8 @@ namespace System.CommandLine.Invocation
         private CancellationTokenSource? _cts;
         private Action<CancellationTokenSource>? _cancellationHandlingAddedEvent;
         private HelpBuilder? _helpBuilder;
+        private BindingContext? _bindingContext;
+        private IConsole _console;
 
         /// <param name="parseResult">The result of the current parse operation.</param>
         /// <param name="console">The console to which output is to be written.</param>
@@ -23,20 +26,32 @@ namespace System.CommandLine.Invocation
             ParseResult parseResult,
             IConsole? console = null)
         {
-            BindingContext = new BindingContext(parseResult, console);
-            BindingContext.ServiceProvider.AddService(_ => GetCancellationToken());
-            BindingContext.ServiceProvider.AddService(_ => this);
+            ParseResult = parseResult;
+            _console = console ?? new SystemConsole();
         }
 
         /// <summary>
         /// The binding context for the current invocation.
         /// </summary>
-        public BindingContext BindingContext { get; }
+        public BindingContext BindingContext
+        {
+            get
+            {
+                if (_bindingContext == null)
+                {
+                    _bindingContext = new BindingContext(ParseResult, Console);
+                    _bindingContext.ServiceProvider.AddService(_ => GetCancellationToken());
+                    _bindingContext.ServiceProvider.AddService(_ => this);
+                }
+
+                return _bindingContext;
+            }
+        }
 
         /// <summary>
         /// The console to which output should be written during the current invocation.
         /// </summary>
-        public IConsole Console => BindingContext.Console;
+        public IConsole Console => _bindingContext?.Console ?? _console;
 
         /// <summary>
         /// Enables writing help output.
@@ -46,7 +61,7 @@ namespace System.CommandLine.Invocation
         /// <summary>
         /// The parser used to create the <see cref="ParseResult"/>.
         /// </summary>
-        public Parser Parser => BindingContext.ParseResult.Parser;
+        public Parser Parser => ParseResult.Parser;
 
         /// <summary>
         /// Provides localizable strings for help and error messages.
@@ -56,11 +71,7 @@ namespace System.CommandLine.Invocation
         /// <summary>
         /// The parse result for the current invocation.
         /// </summary>
-        public ParseResult ParseResult
-        {
-            get => BindingContext.ParseResult;
-            set => BindingContext.ParseResult = value;
-        }
+        public ParseResult ParseResult { get; set; }
 
         /// <summary>
         /// A value that can be used to set the exit code for the process.