… src/System.CommandLine.ApiCompatibility.Tests/ApiCompatibilityApprovalTests.System_CommandLine_api_is_not_changed.approved.txt
… src/System.CommandLine/Parsing/SyntaxVisitor.cs