@jonsequitur Another set of improvements ;)

|        Method |      Args | Mean before | Mean after | Allocated before | Allocated  after|
|-------------- |---------- |------------:|-----------:|-----------------:|----------------:|
|  DefaultsSync | String[0] |   30.004 us |  27.749 us |         14,010 B |        11,850 B |
| DefaultsAsync | String[0] |   29.371 us |  27.667 us |         14,010 B |        11,849 B |
|   MinimalSync | String[0] |    3.366 us |   2.608 us |          6,753 B |         4,961 B |
|  MinimalAsync | String[0] |    3.453 us |   2.780 us |          6,753 B |         4,961 B |
|  DefaultsSync | String[4] |   33.569 us |  32.414 us |         17,432 B |        15,152 B |
| DefaultsAsync | String[4] |   33.727 us |  32.714 us |         17,432 B |        15,154 B |
|   MinimalSync | String[4] |    5.739 us |   4.892 us |         10,176 B |         8,264 B |
|  MinimalAsync | String[4] |    5.888 us |   4.920 us |         10,176 B |         8,264 B |

API changes: after looking at the code I got to the conclusion, that the only `Symbol` which might have multiple children is `Command`, as `Argument` should have none and `Option` might have one: a single `Argument`:

https://github.com/dotnet/command-line-api/blob/74ae46ea8571269a66945049ce9f134a91d67d1e/src/System.CommandLine/Option.cs#L143

2nd biggest gain was allocating `BindingContext` when it's actually needed.